Improve file uploading
Summary: In order to upload large files, we need to use a special package that correctly streams files to the remote server without reading the entire file in to ram. Files larger than 2Gb are impossible to upload without streaming, because it crashes the request package. An additional fix was also added which correctly names the uploaded file in the request. Previously it was simply "null".

Test Plan: TBD

Reviewers: TomReichardt

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D590